Make MainActivity scrollable and add Toppings
Wraps the entire LinearLayout of MainActivity in a ScrollView so that
the activity is scrollable is it becomes too large. Adds CheckBox
views and TextView to represent toppings section of the coffee order.
The displayPrice method has been rewritten to take these options into
account, and adds $1 for whipped cream, and $2 for the chocolate.