Add calculatePrice method and cleanup code
Since we're going to be using the price in other places, it is prudent
now to create a calculatePrice method. Because displayPrice no longer
calculates the price, I changed its first parameter to be the price,
rather than the quantity.  I also took this opportunity to cleanup my
String.format call in displayPrice.