Add OpenWeatherMap HTTP request and ListView
App now requires configuration of MyOpenWeatherMapApiKey variable
with gradle. The developer may set this value within
<HOME>/.gradle/gradle.properties using the format
MyOpenWeatherMapApiKey="" with the key value between the quotation
marks.

I removed the boilerplate FloatingActionButton from MainActvity.java
and activity_mail.xml. I created a basic template view,
list_item_forecast.xml for a single forecast item to be represented in
the main fragment's ListView. I also added a ListView to
fragment_main.xml to represent our list of daily weather forecasts, and
changed its root View to a FrameLayout. Finally, I added an HTTP
request for the OpenWeatherMap API to
MainActivityFragment.onCreateView, which crashes the app with a
NetworkOnMainThreadException, but this will be resolved shortly.