Parse OpenWeatherMap data on background thread
Makes an HTTP request to the OpenWeatherMap API using AsyncTask. Parses
returned JSON data into a String array, and passes this array to the
ArrayAdapter for display within our ListView.
Adds android.permission.INTERNET in order to make this request.

Adds a manual refresh button to the ForecastFragment, which is inflated
within the MainActivity's menu. Moves MainActivityFragment.java to
ForecastFragment.java.