Add and launch forecast detail activity
Deviates a lot from the instructor supplied code due to changes in the
way new activities are created by Android Studio. Seems there was a bug
in my attempt to merge the instructor's code causing my fragment to be
created twice. I blindly replaced their code with the boilerplate from
MainActivity.java, and this seems to have solved that problem.

Along with this change, I also have a content_NN.xml for every new
activity, which notably bloats the size and complexity of the project.
This is apparently what Android Studio wants to do, as I have not
discovered another option to create a blank activity with a fragment.

The expected activity hierarchy is disrupted by this as I don't get a
back arrow on my DetailsActivity.  I'm just going to push through it.